refactoring auth module into several files
