+++ functions/client/src/store/modules/Auth/index.js
@@ -1,192 +1,7 @@
-import api from '../../../api/index';
-import router from '../../../router/index';
-import * as firebase from 'firebase';
-
-const state = {
-  user: null,
-  error: null,
-  spotifyAuthAccessCode: window.localStorage.getItem('spotifyAuthAccessCode'),
-  spotifyAuthRefreshCode: window.localStorage.getItem('spotifyAuthRefreshCode')
-};
-const getters = {
-  isSpotifyLoggedIn: state => !!state.spotifyAuthAccessCode,
-  user: state => state.user,
-  error: state => state.error,
-  getAccessToken: state => state.spotifyAuthAccessCode,
-  getRefreshToken: state => state.spotifyAuthRefreshCode
-};
-const actions = {
-  firebaseSignUpUser({ commit, dispatch }, signUpRequest) {
-    commit('SET_LOADING', true);
-    commit('clearError');
-    firebase
-      .auth()
-      .createUserWithEmailAndPassword(
-        signUpRequest.email,
-        signUpRequest.password
-      )
-      .then(response => {
-        commit('SET_LOADING', false);
-        const newUser = {
-          id: response.user.uid
-        };
-        commit('setUser', newUser);
-
-        if (signUpRequest.linkSpotify) {
-          window.localStorage.setItem('fbUserId', newUser.id);
-          dispatch('loginSpotify');
-        }
-      })
-      .catch(error => {
-        commit('SET_LOADING', false);
-        commit('setError', error);
-        console.log(error);
-      });
-  },
-  firebaseSignInUser({ commit, dispatch }, signInRequest) {
-    commit('SET_LOADING', true);
-    commit('clearError');
-    firebase
-      .auth()
-      .signInWithEmailAndPassword(signInRequest.email, signInRequest.password)
-      .then(response => {
-        commit('SET_LOADING', false);
-        const oldUser = {
-          id: response.user.uid
-        };
-        commit('setUser', oldUser);
-        dispatch('fetchUserSpotAuthTokenFromFB', oldUser.id);
-      })
-      .catch(error => {
-        commit('SET_LOADING', false);
-        commit('setError', error);
-        console.log(error);
-      });
-  },
-  autoSignIn({ commit }, payload) {
-    const user = {
-      id: payload.uid
-    };
-    commit('setUser', user);
-  },
-  logout({ commit, getters, dispatch }) {
-    if (getters.isSpotifyLoggedIn) {
-      dispatch('logoutSpotify');
-    }
-    firebase.auth().signOut();
-    commit('setUser', null);
-
-    router.push('/landing');
-  },
-  clearError({ commit }) {
-    commit('clearError');
-  },
-  loginSpotify: () => {
-    api.loginSpotify();
-  },
-
-  finalizeSpotifyLogin({ commit }, tokens) {
-    commit('setSpotifyAuthCodes', tokens);
-    window.localStorage.setItem('spotifyAuthAccessCode', tokens.access_token);
-    window.localStorage.setItem('spotifyAuthRefreshCode', tokens.refresh_token);
-
-    router.push('/saved/playlists');
-  },
-
-  saveUserSpotAuthTokensToFB({ commit, dispatch }, query) {
-    const userId = window.localStorage.getItem('fbUserId');
-
-    const tokens = {
-      access_token: query.access_token,
-      refresh_token: query.refresh_token
-    };
-
-    commit('SET_LOADING', true);
-
-    firebase
-      .database()
-      .ref('/users/' + userId)
-      .child('/tokens/')
-      .push(tokens)
-      .then(() => {
-        dispatch('finalizeSpotifyLogin', tokens);
-      })
-      .catch(err => {
-        console.log(err.message);
-      })
-      .finally(() => {
-        commit('SET_LOADING', false);
-      });
-  },
-
-  fetchUserSpotAuthTokenFromFB({ dispatch }, id) {
-    firebase
-      .database()
-      .ref('/users/' + id)
-      .child('/tokens/')
-      .once('value')
-      .then(data => {
-        let tokens = {
-          access_token: null,
-          refresh_token: null
-        };
-        const obj = data.val();
-        for (let key in obj) {
-          tokens.access_token = obj[key].access_token;
-          tokens.refresh_token = obj[key].refresh_token;
-        }
-
-        dispatch('finalizeSpotifyLogin', tokens);
-      });
-  },
-
-  logoutSpotify({ commit }) {
-    commit('setSpotifyAuthCodes', {
-      access_token: null,
-      refresh_token: null
-    });
-
-    window.localStorage.removeItem('spotifyAuthAccessCode');
-    window.localStorage.removeItem('spotifyAuthRefreshCode');
-    window.localStorage.removeItem('fbUserId');
-  },
-  fetchSpotifyRefreshToken({ commit, getters }) {
-    if (!getters.isSpotifyLoggedIn) return;
-    api
-      .fetchSpotifyRefreshToken(getters.getRefreshToken)
-      .then(({ data }) => {
-        commit('setSpotifyAuthCodes', {
-          access_token: data.items.access_token,
-          refresh_token: data.items.refresh_token
-        });
-        window.localStorage.setItem(
-          'spotifyAuthAccessCode',
-          data.items.access_token
-        );
-        window.localStorage.setItem(
-          'spotifyAuthRefreshCode',
-          data.items.refresh_token
-        );
-        return Promise.resolve(true);
-      })
-      .catch(err => console.log(err.message));
-  }
-};
-const mutations = {
-  setUser: (state, userObject) => {
-    state.user = userObject;
-  },
-  setError: (state, error) => {
-    state.error = error;
-  },
-  clearError(state) {
-    state.error = null;
-  },
-  setSpotifyAuthCodes: (state, { access_token, refresh_token }) => {
-    state.spotifyAuthAccessCode = access_token;
-    state.spotifyAuthRefreshCode = refresh_token;
-  }
-};
+import state from './state';
+import getters from './getters';
+import actions from './actions';
+import mutations from './mutations';
 
 export default {
   state,
