+++ functions/client/src/store/modules/Auth/state.js
@@ -1,0 +1,8 @@
+const state = {
+  user: null,
+  error: null,
+  spotifyAuthAccessCode: window.localStorage.getItem('spotifyAuthAccessCode'),
+  spotifyAuthRefreshCode: window.localStorage.getItem('spotifyAuthRefreshCode')
+};
+
+export default state;
